pyproject.toml
[project]
name = "ifclient"
version = "0.1.0"
description = "A CLI tool to configure and change InsightFinder project settings"
authors = [
    {name = "Raja Shekar Reddy Siriganagari",email = "raja@insightfinder.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "click",
    "requests",
    "pydantic",
    "pyyaml"
]
packages = [
    { include = "ifclient", from = "src" }
]

[tool.poetry.scripts]
ifclient = "ifclient.cli:cli"


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
